Refactor
Renamed Index, Index2, and Index3.html to Groups, Skills, and Users.html respectively.  Added Signoffs.html.  Updated menu system.  Added anchors to each table for the Index page.  Added role checking for pages that require it.  Added a checkbox for password updates when editing a user.  Also added the ability to change the email address for a user and have that change stored in the database as well as updating the list of allowed logins on the fly.  Email address is used as the user name, so this change must be reflected in the users map.